+++ internal/storage/v1/cassandra/spanstore/dbmodel/test_fixtures_test.go
@@ -71,3 +71,26 @@ var (
 		},
 	}
 )
+
+func getTestSpan() *Span {
+	span := &Span{
+		TraceID:       someDBTraceID,
+		SpanID:        someSpanID,
+		OperationName: someOperationName,
+		Flags:         someFlags,
+		StartTime:     int64(model.TimeAsEpochMicroseconds(someStartTime)),
+		Duration:      int64(model.DurationAsMicroseconds(someDuration)),
+		Tags:          someDBTags,
+		Logs:          someDBLogs,
+		Refs:          someDBRefs,
+		Process:       someDBProcess,
+		ServiceName:   someServiceName,
+	}
+	// there is no way to validate if the hash code is "correct" or not,
+	// other than comparing it with some magic number that keeps changing
+	// as the model changes. So let's just make sure the code is being
+	// calculated during the conversion.
+	spanHash, _ := model.HashCode(span)
+	span.SpanHash = int64(spanHash)
+	return span
+}
